+++ korr/__init__.py
@@ -8,3 +8,4 @@ from .corr_vs_pval import corr_vs_pval
 from .bracket_pval import bracket_pval
 from .confusion import confusion
 from .mcc import mcc, confusion_to_mcc
+from .slice_xy import slice_xy